LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/json-module/utf8-expected.txt

CONSOLE MESSAGE: line 1: TypeError: 'text/json' is not a valid JavaScript MIME type.
CONSOLE MESSAGE: line 1: TypeError: 'text/json' is not a valid JavaScript MIME type.
CONSOLE MESSAGE: line 1: TypeError: 'text/json' is not a valid JavaScript MIME type.
CONSOLE MESSAGE: line 1: TypeError: 'text/json' is not a valid JavaScript MIME type.

FAIL utf-8 assert_unreached: [object Event] Reached unreachable code
FAIL shift-jis assert_unreached: [object Event] Reached unreachable code
FAIL windows-1252 assert_unreached: [object Event] Reached unreachable code
FAIL utf-7 assert_unreached: [object Event] Reached unreachable code
